+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## v1.0.105
+#### **firehose-metrics**
+### 💡 Enhancements 
+- Added custom naming for global resources
+- Added ability to import global resources (s3 & iam)
+### 🛑 Breaking changes 🛑
+- Update variables: `private_key` renamed to `api_key` with type `string` instead of `any`.
+
 ## v1.0.104
 #### **msk-data-stream**
 ### 🚀 New module 🚀